requirements.txt
fastapi
uvicorn[standard]
pyswip
python-dotenv